@@ -2,7 +2,7 @@
 name = "gopher64"
 version = "1.0.18"
 edition = "2024"
-rust-version = "1.87"
+rust-version = "1.88"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
